clean code, addressed comments, fixed tests
